srv-django-backend/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=21.2.0
dj-database-url>=2.1.0
django-jazzmin==2.6.0
whitenoise==6.6.0
django-extensions==3.2.3
werkzeug==3.0.1
pyOpenSSL==24.0.0
django-import-export>=3.3.0
openpyxl>=3.1.0